To pass data to a function we define parameters which allows us to give input to function and use that input for performing some task.
We don't have to declare a variable when it's used as a parameter in a
function definition. When it's used as a parameter, the declaration is
implicit.

There is no such rule that the argument and the name of parameter should
match

Argument is the variable or any object which is passed to function as a
parameter

A function can have any numbers of parameters seperated by commas