We can change existing image with another image in HTML with the help of js
In this code

OnMouseover Event chngPic method is called.
The method require two arguments one is element id and second is source
of the image to be replace
Method on execution target image tag by document.getElementById getting
the id from method arguments
src property is assigned with the source argument of chngPic method
so on getting the mouse over an image will change the existing image.