JS main purpose was to create interactivity between a user and website and one of the most interative feature of web is Forms where a website get information of a visitor but what about information authenticness.
JS allows us to control and valid user input in a form so a website owner
can get correct information and user should get idea what sort of
information is a valid for the form.

In this code we are validating input text field of a form, the code checks
that input text field must contain some text otherwise it will show alert
that the input field is empty or null and it is required to fill.

HTML5 required attribute do the same thing.