try catch is used tackle error in JS code
The main code is written in try code block i.e. try to execute this code
if the code the code produce error then this is determine in catch block
i.e. catch the error

if code produce error this error is assigned to catch as a parameter
and can be displayed to user as a easy non-technical message by alert or
any other way or by displaying a default message which is recieved in
catch parameter by JS